refactor: Delegate tracking logic to SimpleTracker and decouple CLI
