+++ src/pystormtracker/models/__init__.py
@@ -1,6 +1,7 @@
 from .center import Center, DetectedCenters
 from .grid import Grid
 from .time import TimeRange
+from .tracker import Tracker
 from .tracks import Tracks
 
-__all__ = ["Center", "DetectedCenters", "Grid", "TimeRange", "Tracks"]
+__all__ = ["Center", "DetectedCenters", "Grid", "TimeRange", "Tracker", "Tracks"]
