+++ src/pystormtracker/simple/__init__.py
@@ -1,4 +1,5 @@
 from .detector import SimpleDetector
 from .linker import SimpleLinker
+from .tracker import SimpleTracker
 
-__all__ = ["SimpleDetector", "SimpleLinker"]
+__all__ = ["SimpleDetector", "SimpleLinker", "SimpleTracker"]
